atlas/ghost-shift-hunter-final/backend/requirements.txt
fastapi
uvicorn
pydantic
openai
pandas
thefuzz
python-levenshtein
python-multipart
pydantic-ai
google-generativeai
